Refactor - Registered new case file component Registered new case file component's view with the router Changed import to point to new case file Renamed and moved around files
