+++ src/app.container.ts
@@ -1,18 +1,18 @@
 import 'reflect-metadata'
 import { container } from 'inversify-props'
 import IHomeDashboardService from '@/services/interfaces/IHomeDashboardService'
-import IAddonsService from './services/interfaces/IAddonsService'
-import ICaseFileService from './services/interfaces/ICaseFileService'
-import INetworkMonitorService from './services/interfaces/INetworkMonitorService'
+import IAddonsService from '@/services/interfaces/IAddonsService'
+import ICaseFileService from '@/components/case-file/services/ICaseFileService'
+import INetworkMonitorService from '@/services/interfaces/INetworkMonitorService'
 import ICameraMonitorService from '@/services/interfaces/ICameraMonitorService'
 import AddonsService from '@/services/implementations/AddonsService'
-import CaseFileService from './services/implementations/CaseFileService'
-import HomeDashboardService from './services/implementations/HomeDashboardService'
-import NetworkMonitorService from './services/implementations/NetworkMonitorService'
+import CaseFileService from '@/components/case-file/services/CaseFileService'
+import HomeDashboardService from '@/services/implementations/HomeDashboardService'
+import NetworkMonitorService from '@/services/implementations/NetworkMonitorService'
 import TYPES from '@/InjectableTypes/types'
 import IMenuItemService from '@/services/interfaces/IMenuItemService'
 import MenuItemService from '@/services/implementations/MenuItemService'
-import CameraMonitorService from './services/implementations/CameraMonitorService'
+import CameraMonitorService from '@/services/implementations/CameraMonitorService'
 
 export default function registerContainerServices () {
   container.options.skipBaseClassChecks = true
