+++ src/components/NavDrawer.vue
@@ -68,7 +68,7 @@ import HelloWorld from '@/components/HelloWorld.vue'
 import IMenuItem from '@/types/IMenuItem'
 import IHomeDashboardService from '@/services/interfaces/IHomeDashboardService'
 import IAddonsService from '@/services/interfaces/IAddonsService'
-import ICaseFileService from '@/services/interfaces/ICaseFileService'
+import ICaseFileService from './case-file/services/ICaseFileService'
 import INetworkMonitorService from '@/services/interfaces/INetworkMonitorService'
 import ICameraMonitorService from '@/services/interfaces/ICameraMonitorService'
 import TYPES from '@/InjectableTypes/types'
