+++ src/components/add-on-manager/AddOnsCard.vue
