+++ src/components/case-file/services/ICaseFileService.ts
@@ -1,4 +1,4 @@
-import IService from './IService'
+import IService from '@/services/interfaces/IService'
 
 export default interface ICaseFileService extends IService {
   getCaseFileData() : object
