+++ src/router/index.ts
@@ -1,8 +1,8 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import Home from '@/views/HomeView.vue'
-import CaseFile from '@/views/CaseFileView.vue'
-import CaseFileReport from '@/views/CaseFileReportView.vue'
+import CaseFile from '@/components/case-file/CaseFileView.vue'
+import CaseFileReport from '@/components/case-file/CaseFileReportView.vue'
 import NetworkMonitor from '@/views/NetworkMonitorView.vue'
 import AddOns from '@/views/AddOnsView.vue'
 import CameraMonitor from '@/views/CameraMonitorView.vue'
